Implement high-level topic producer
Core client is responsible for managing the lifecycle of cluster
bootstrapping, API versions, metadata discovery, connecting to new
brokers, reconnecting upon transient errors, and so on.

The core client is a series of tasks which cooperate together to provide
a uniform, topic-oriented interface to the backing Kafka cluster.

For every topic that an app needs to publish to, a TopicProducer may be
created which multiplexes all of its broker interaction over the same
TCP connections of the core client.

The TopicProducer exclusively uses the uniform sticky producer strategy,
for performance, atomicity, efficiency, and simplicity. The target
partition is rotated after every interaction to ensure distribution of
load across partitions.